refactor: moves sorting algos into own folder
